doOutput() in doRest() to modify its pointer only in safe ways. (#20)
Without this change, JSON structs can change in strange ways after
assignment. This is because when the go json package unmarshals, it
modified slices and maps within a struct in place. The way to fix this
is to set the struct to its zero value before unmarshalling. Since the
zero value has no allocated slices or maps, the json library will
allocate new slices and maps rather than modifying existing ones in
place which is what we need for the assignment operator to work as
intended.

Co-authored-by: Travis Keep <travisk@vmware.com>